+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.29.18] - 2021-10-13
+### Fixed
+- Fixes for rename to nested directory
+- Renaming only if shadow exists
+- fs.rmSync -> fs.unlinkSync
+
 ## [2.29.17] - 2021-10-13
 ### Added
 - Added unit tests for detectBranchChange